Move `allow_access` to `gateway.rs`
