+++ rust/connlib/tunnel/src/control_protocol/gateway.rs
@@ -116,47 +116,6 @@ where
         })
     }
 
-    pub fn allow_access(
-        &mut self,
-        resource: ResourceDescription,
-        client: ClientId,
-        expires_at: Option<DateTime<Utc>>,
-        domain: Option<Dname>,
-    ) -> Option<DomainResponse> {
-        let peer = self.role_state.peers.get_mut(&client)?;
-
-        let (addresses, resource_id) = match &resource {
-            ResourceDescription::Dns(r) => {
-                let Some(domain) = domain.clone() else {
-                    return None;
-                };
-
-                if !is_subdomain(&domain, &r.domain) {
-                    return None;
-                }
-
-                (r.addresses.clone(), r.id)
-            }
-            ResourceDescription::Cidr(cidr) => (vec![cidr.address], cidr.id),
-        };
-
-        for address in &addresses {
-            peer.transform
-                .add_resource(*address, resource.clone(), expires_at);
-        }
-
-        tracing::info!(%client, resource = %resource_id, expires = ?expires_at.map(|e| e.to_rfc3339()), "Allowing access to resource");
-
-        if let Some(domain) = domain {
-            return Some(DomainResponse {
-                domain,
-                address: addresses.iter().map(|i| i.network_address()).collect(),
-            });
-        }
-
-        None
-    }
-
     fn new_peer(
         &mut self,
         ips: Vec<IpNetwork>,
