+++ rust/connlib/tunnel/src/gateway.rs
@@ -1,13 +1,17 @@
 use std::task::{ready, Context, Poll};
 use std::time::Duration;
 
+use crate::control_protocol::gateway::ResourceDescription;
 use crate::device_channel::Device;
 use crate::ip_packet::MutableIpPacket;
 use crate::peer::PacketTransformGateway;
 use crate::peer_store::PeerStore;
 use crate::Tunnel;
-use connlib_shared::messages::{ClientId, Interface as InterfaceConfig, ResourceId};
-use connlib_shared::Callbacks;
+use chrono::{DateTime, Utc};
+use connlib_shared::messages::{
+    ClientId, DomainResponse, Interface as InterfaceConfig, ResourceId,
+};
+use connlib_shared::{Callbacks, Dname};
 use snownet::Server;
 use tokio::time::{interval, Interval, MissedTickBehavior};
 
@@ -41,6 +45,17 @@ where
     /// Clean up a connection to a resource.
     pub fn cleanup_connection(&mut self, id: &ClientId) {
         self.role_state.peers.remove(id);
+    }
+
+    pub fn allow_access(
+        &mut self,
+        resource: ResourceDescription,
+        client: ClientId,
+        expires_at: Option<DateTime<Utc>>,
+        domain: Option<Dname>,
+    ) -> Option<DomainResponse> {
+        self.role_state
+            .allow_access(resource, client, expires_at, domain)
     }
 
     pub fn remove_access(&mut self, id: &ClientId, resource_id: &ResourceId) {
@@ -78,6 +93,47 @@ impl GatewayState {
         ready!(self.expire_interval.poll_tick(cx));
         self.expire_resources();
         Poll::Ready(())
+    }
+
+    pub fn allow_access(
+        &mut self,
+        resource: ResourceDescription,
+        client: ClientId,
+        expires_at: Option<DateTime<Utc>>,
+        domain: Option<Dname>,
+    ) -> Option<DomainResponse> {
+        let peer = self.peers.get_mut(&client)?;
+
+        let (addresses, resource_id) = match &resource {
+            ResourceDescription::Dns(r) => {
+                let Some(domain) = domain.clone() else {
+                    return None;
+                };
+
+                if !crate::dns::is_subdomain(&domain, &r.domain) {
+                    return None;
+                }
+
+                (r.addresses.clone(), r.id)
+            }
+            ResourceDescription::Cidr(cidr) => (vec![cidr.address], cidr.id),
+        };
+
+        for address in &addresses {
+            peer.transform
+                .add_resource(*address, resource.clone(), expires_at);
+        }
+
+        tracing::info!(%client, resource = %resource_id, expires = ?expires_at.map(|e| e.to_rfc3339()), "Allowing access to resource");
+
+        if let Some(domain) = domain {
+            return Some(DomainResponse {
+                domain,
+                address: addresses.iter().map(|i| i.network_address()).collect(),
+            });
+        }
+
+        None
     }
 
     fn expire_resources(&mut self) {
